+++ src/frontend/apps/impress/src/features/docs/docs-grid/components/DocGridContentList.tsx
@@ -23,9 +23,11 @@ import { Droppable } from './Droppable';
 
 const ModalConfirmationMoveDoc = dynamic(
   () =>
-    import('./ModalConfimationMoveDoc').then((mod) => ({
-      default: mod.ModalConfirmationMoveDoc,
-    })),
+    import('@/docs/doc-management/components/ModalConfimationMoveDoc').then(
+      (mod) => ({
+        default: mod.ModalConfirmationMoveDoc,
+      }),
+    ),
   { ssr: false },
 );